Flow improvement - Addition of CDAI.  (#7)
* Update from V5_FullScript.R to V6_FullScript.R

Added new comments per definitions. Including a sensitivity analysis that replaces the DAS28 with the CDAI. 

- Need to add defined ggsave for each figure in next iteration.

* Update V5_D2TDefinitions

Inclusion of new sensitivity analysis with CDAI 

Inclusion of clearer description per definition

* Adding CDAI

Creation of the CDAI score after imputation

* Replaced more prominent sensivity

Included the CDAI sensivity measure and removed (with #) the previous sensivity that removes the inclusion of a 3rd MOA as a change in disease activity, due to the high similarities. 
It can be added again if preferred by removing the #

* Update Table1 Summary with treatment information

---------

Co-authored-by: ChiaraRipepi <c.ripepi2@umcutrecht.nl>